feat(mis): refactored changePasswordModal
